+++ src/index.js
@@ -105,6 +105,31 @@ function createProgram() {
       'portal token (or APIDEX_TOKEN)',
       process.env.APIDEX_TOKEN,
     )
+    .option(
+      '--clientId <clientId>',
+      'add the clientId from your OpenID Connect provider linked to the developer portal',
+      process.env.APIDEX_CLIENTID,
+    )
+    .option(
+      '--clientSecret <clientSecret>',
+      'add the clientSecret from your OpenID Connect provider linked to the developer portal',
+      process.env.APIDEX_SECRET,
+    )
+    .option(
+      '--aud <aud>',
+      'Only used in combination with client certificate authentication instead of clientSecret. Provide the audience for the client token.',
+      null,
+    )
+    .option(
+      '--scope <scope>',
+      'add the scope for the developer portal app registration',
+      process.env.APIDEX_SCOPE,
+    )
+    .option(
+      '--tokenUrl <tokenUrl>',
+      'add the tokenUrl from your OpenID Connect provider (ex: https://login.microsoftonline.com/yourcompany.onmicrosoft.com/oauth2/v2.0/token)',
+      process.env.APIDEX_TOKENURL,
+    )
     .description(
       'validate OpenAPI overlay files and that manifest API products exist in the portal',
     )
@@ -115,6 +140,11 @@ function createProgram() {
         host: command.host,
         environment: command.environment,
         token: command.token,
+        clientId: command.clientId,
+        clientSecret: command.clientSecret,
+        aud: command.aud,
+        scope: command.scope,
+        tokenUrl: command.tokenUrl,
       });
     });